add interfaces for augmenting sizes and variants
