+++ packages/bootstrap-vue-3/src/types/TextColorVariant.d.ts
@@ -1,0 +1,17 @@
+import {BaseColorVariant} from '.'
+
+/**
+ * @external
+ */
+ export interface BaseTextColorVariant extends BaseColorVariant {
+  'white': unknown
+  'body': unknown
+  'muted': unknown
+  'black-50': unknown
+  'white-50': unknown
+  'reset': unknown
+}
+
+type TextColorVariant = keyof BaseTextColorVariant
+
+export default TextColorVariant
