+++ packages/bootstrap-vue-3/src/types/index.ts
@@ -12,9 +12,9 @@ export type {
   RowColsBreakpointProps,
 } from './BreakpointProps'
 export type {default as ButtonType} from './ButtonType'
-export type {default as ButtonVariant} from './ButtonVariant'
+export type {default as ButtonVariant, BaseButtonVariant} from './ButtonVariant'
 export type {default as ClassValue} from './ClassValue'
-export * from './ColorVariant'
+export type {default as ColorVariant, BaseColorVariant} from './ColorVariant'
 export type {ContainerHorizontalAlign, ContainerPosition, ContainerVerticalAlign} from './Container'
 export type {default as InputSize} from './InputSize'
 export type {default as InputType} from './InputType'
@@ -26,7 +26,6 @@ export type {default as PlaceholderAnimation} from './PlaceholderAnimation'
 export type {default as PlaceholderSize} from './PlaceholderSize'
 export type {default as Position} from './Position'
 export type {default as RadioOption} from './RadioOption'
-export * as RoundedTypes from './RoundedTypes'
 export type {default as SelectOption} from './SelectOption'
 export type {default as Size, BaseSize} from './Size'
 export type {default as SkeletonAnimation} from './SkeletonAnimation'
@@ -34,5 +33,6 @@ export type {default as SkeletonType} from './SkeletonType'
 export type {default as SpinnerType} from './SpinnerType'
 export type {default as TableField, TableFieldObject} from './TableField'
 export type {default as TableItem} from './TableItem'
+export type {default as TextColorVariant, BaseTextColorVariant} from './TextColorVariant'
 export type {default as TransitionMode} from './TransitionMode'
 export type {default as VerticalAlign} from './VerticalAlign'
